change MainPage
