+++ apps/frontend/src/App.tsx
@@ -1,14 +1,19 @@
+import { useEffect } from 'react'
 import { Layout, theme } from 'antd'
-import { Breadcrumb } from './components/Breadcrumb'
+import { microserviceData } from "./store"
+
 import { Pages } from './pages'
-import { HeaderMenu } from './components/HeaderMenu'
-import { Logo } from './components/Logo'
+import { Logo, HeaderMenu, Breadcrumb } from './components'
 
 import './App.css'
 
 const { Header, Content, Footer } = Layout
 
 const App: React.FC = () => {
+    useEffect(() => {
+        microserviceData.fetchData()
+    }, [])
+
     const {
         token: { colorBgContainer },
     } = theme.useToken()
