--- apps/frontend/src/components/MicroserviceListInfo/index.ts
@@ -1,1 +1,0 @@
-export { MicroserviceListInfo } from './MicroserviceListInfo'
