+++ apps/frontend/src/components/index.ts
@@ -1,0 +1,4 @@
+export { Breadcrumb } from './Breadcrumb'
+export { HeaderMenu } from './HeaderMenu'
+export { HealthStatus } from './HealthStatus'
+export { Logo } from './Logo'
