+++ apps/frontend/src/pages/MainPage/MainPage.module.css
