+++ apps/frontend/src/pages/MainPage/MainPage.tsx
@@ -1,6 +1,10 @@
 import { observer } from "mobx-react"
 import { MonitoredMicroservice } from "@nats-micro-monitor/types"
-import { MicroserviceListInfo } from "../../components/MicroserviceListInfo"
+
+import { List } from "./List"
+import { Header } from "./Header"
+
+import styles from './MainPage.module.css'
 
 interface Props {
     data: MonitoredMicroservice[]
@@ -9,7 +13,13 @@ interface Props {
 export const MainPage = observer(({ data }: Props) => {
     return (
         <div>
-            <MicroserviceListInfo items={data} />
+            <table className={styles.Container}>
+                <Header />
+
+                <tbody className={styles.Content}>
+                    <List items={data} />
+                </tbody>
+            </table>
         </div>
     )
 })
