+++ apps/frontend/src/pages/StatusPage/index.ts
@@ -1,0 +1,1 @@
+export { StatusPage } from './StatusPage'
